requirements.txt
fastapi==0.115.0
pydantic==2.9.2
python-dotenv==1.0.1
supabase==2.8.1
schedule==1.2.2
